+++ frontend/src/component/feature/FeatureView/FeatureOverview/FeatureOverviewMetaData/FeatureOverviewSidePanelDetails.test.tsx
@@ -1,10 +1,16 @@
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { render } from 'utils/testRenderer';
+import { render } from '../../../../../utils/testRenderer';
 import { FeatureOverviewSidePanelDetails } from './FeatureOverviewSidePanelDetails';
-import type { IDependency, IFeatureToggle } from 'interfaces/featureToggle';
-import { testServerRoute, testServerSetup } from 'utils/testServer';
-import ToastRenderer from 'component/common/ToastRenderer/ToastRenderer';
+import type {
+    IDependency,
+    IFeatureToggle,
+} from '../../../../../interfaces/featureToggle';
+import {
+    testServerRoute,
+    testServerSetup,
+} from '../../../../../utils/testServer';
+import ToastRenderer from '../../../../common/ToastRenderer/ToastRenderer';
 
 const server = testServerSetup();
