+++ frontend/src/component/feature/FeatureView/FeatureOverview/FeatureOverviewMetaData/VariantsTooltip.tsx
@@ -1,4 +1,4 @@
-import { TooltipLink } from 'component/common/TooltipLink/TooltipLink';
+import { TooltipLink } from '../../../../common/TooltipLink/TooltipLink';
 import type { FC } from 'react';
 
 export const VariantsTooltip: FC<{
