+++ frontend/src/component/feature/FeatureView/FeatureOverview/FeatureOverviewMetaData/useShowDependentFeatures.ts
@@ -1,5 +1,5 @@
-import { useCheckDependenciesExist } from 'hooks/api/getters/useCheckDependenciesExist/useCheckDependenciesExist';
-import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
+import { useCheckDependenciesExist } from '../../../../../hooks/api/getters/useCheckDependenciesExist/useCheckDependenciesExist';
+import useUiConfig from '../../../../../hooks/api/getters/useUiConfig/useUiConfig';
 
 export const useShowDependentFeatures = (project: string) => {
     const { dependenciesExist } = useCheckDependenciesExist(project);
